# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Text editor MVP API
  version: 1.0.0
  description: Identity, Document and Processing contracts; ownership, folder and
    save semantics in docs.
servers:
  - url: http://localhost:8080
    description: Local same-origin reverse proxy
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/documents/{documentId}/permissions:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: listPermissions
      summary: Owner list explicit grants
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionPage'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - *id003
        - *id004
    post:
      tags:
        - Document
      x-service: document-service
      operationId: grantByEmail
      summary: Owner grant existing account by email
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantEmailRequest'
  /api/v1/documents/{documentId}/permissions/{granteeUserId}:
    put:
      tags:
        - Document
      x-service: document-service
      operationId: putPermission
      summary: Owner create or change explicit grant
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - name: granteeUserId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantRequest'
    delete:
      tags:
        - Document
      x-service: document-service
      operationId: deletePermission
      summary: Owner revoke grant idempotently
      responses:
        '204':
          description: Success with no body.
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - name: granteeUserId
          in: path
          required: true
          schema: *id001
  /api/v1/documents/{documentId}/share-links:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: listShareLinks
      summary: Owner list links without raw tokens
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkPage'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - *id003
        - *id004
    post:
      tags:
        - Document
      x-service: document-service
      operationId: createShareLink
      summary: Owner create expiring read-only link
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedShareLink'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareLinkRequest'
  /api/v1/documents/{documentId}/share-links/{linkId}:
    delete:
      tags:
        - Document
      x-service: document-service
      operationId: revokeShareLink
      summary: Owner revoke link idempotently
      responses:
        '204':
          description: Success with no body.
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - name: linkId
          in: path
          required: true
          schema: *id001
  /api/v1/public/shares/{token}:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: readPublicShare
      summary: Read public metadata without owner or folder
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocument'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 43
            pattern: ^[A-Za-z0-9_-]{43}$
            maxLength: 43
      security: []
  /api/v1/public/shares/{token}/content:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: readPublicContent
      summary: Stream current native snapshot after token validation
      responses:
        '200':
          description: Native snapshot bytes; no public storage URL.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
        '204':
          description: Virtual empty document, no snapshot yet.
      description: Invalid/revoked/expired/trash =>404. No storage URL. Empty head
        revision0 returns204; frontend metadata empty=true.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 43
            pattern: ^[A-Za-z0-9_-]{43}$
            maxLength: 43
      security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          maxLength: 64
        message:
          type: string
          maxLength: 500
        traceId:
          type: string
          maxLength: 64
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
      required:
        - code
        - message
        - traceId
...
    Permission:
      type: object
      properties:
        granteeUserId: *id001
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          maxLength: 100
        role:
          type: string
          enum:
            - VIEWER
            - EDITOR
        updatedAt: *id006
      additionalProperties: false
      required:
        - granteeUserId
        - role
        - updatedAt
    GrantRequest:
      type: object
      properties:
        role:
          type: string
          enum:
            - VIEWER
            - EDITOR
      additionalProperties: false
      required:
        - role
    GrantEmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        role:
          type: string
          enum:
            - VIEWER
            - EDITOR
      additionalProperties: false
      required:
        - email
        - role
    ShareLink:
      type: object
      properties:
        id: *id001
        createdAt: *id006
        expiresAt: *id006
        revokedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
      required:
        - id
        - createdAt
        - expiresAt
        - revokedAt
    CreateShareLinkRequest:
      type: object
      properties:
        expiresInSeconds:
          type: integer
          minimum: 3600
          maximum: 2592000
          default: 604800
      additionalProperties: false
    CreatedShareLink:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/ShareLink'
        token:
          type: string
          minLength: 43
          maxLength: 43
        viewerPath:
          type: string
          maxLength: 100
      additionalProperties: false
      required:
        - link
        - token
        - viewerPath
    PublicDocument:
      type: object
      properties:
        documentId: *id001
        title:
          type: string
          maxLength: 200
        headRevision: *id008
        empty:
          type: boolean
        contentPath:
          type: string
          nullable: true
          maxLength: 150
      additionalProperties: false
      required:
        - documentId
        - title
        - headRevision
        - empty
        - contentPath
...
    PermissionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        nextCursor:
          type: string
          nullable: true
      additionalProperties: false
      required:
        - items
        - nextCursor
    ShareLinkPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShareLink'
        nextCursor:
          type: string
          nullable: true
      additionalProperties: false
      required:
        - items
        - nextCursor
...
  responses:
    ErrorResponse:
      description: Error with stable code; semantics and status in docs/05_SAVE_AND_SHARE.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'